# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Robolucha API",
        "title": "Robolucha API",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "http://local.robolucha.com:5000",
    "basePath": "/",
    "paths": {
...
        "/private/media": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "add media",
                "parameters": [
                    {
                        "description": "MediaRequest",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.MediaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Media"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Media": {
            "type": "object",
            "properties": {
                "fileName": {
                    "type": "string"
                },
                "gameDefinition": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "narrativeDefinition": {
                    "type": "integer"
                },
                "thumbnail": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "model.MediaRequest": {
            "type": "object",
            "properties": {
                "base64Data": {
                    "type": "string"
                },
                "fileName": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}